ya va el delete de camisetas :)

+++ src/main/java/com/example/proyectoCamisetas/entity/Categoria.java
@@ -7,11 +7,12 @@ import lombok.NoArgsConstructor;
 @Entity
 @Data
 @NoArgsConstructor
-@Table(name = "categoria")
+@Table(name = "categoria") 
 public class Categoria {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
+    private Integer id; 
+
     private String nombre;
 }
